Refactor after mentor review
