+++ src/ua/javarush/constans/Constans.java
