+++ src/ua/javarush/constans/EncryptionCommandTypes.java
