+++ src/ua/javarush/encryption/Main.java
@@ -1,0 +1,8 @@
+package ua.javarush.encryption;
+
+public class Main {
+    public static void main(String[] args)  {
+        Runner  runner = new Runner();
+        runner.start(args);
+    }
+}
